Changed: BaseField does not forgo rendering anymore
Explanation: If the wrapped component declared e.g. label as a required
prop, the baseField wrapper would forgo the rendering of the label.
This mechanism is not really explicit and caused repeatedly problems.
If that functionality is needed in the future, we can reintroduce it
with a more explicit solutions (such as, the wrapped component must
declare some specific static prop such as requestRenderingFor = {…} )